Add connect functionality
