+++ components/Layout/NavBar/index.tsx
@@ -1,5 +1,8 @@
 import { useState } from "react";
+import { useWeb3React } from "@web3-react/core";
+import { Web3Provider } from "@ethersproject/providers";
 import { AbstractConnector } from "@web3-react/abstract-connector";
+import { useEagerConnect, useInactiveListener } from "../../../hooks";
 
 // Mui
 import { default as MuiAppBar } from "@mui/material/AppBar";
@@ -11,11 +14,23 @@ import Button from "@mui/material/Button";
 // Components
 import DialogWalletOptions from "./DialogWalletOptions";
 
+// Utils
+import getErrorMessage from "../../../utils/getErrorMessage";
+
 export default function AppBar() {
+  const { activate, account, error } = useWeb3React<Web3Provider>();
+  if (!!error) console.log(getErrorMessage(error));
+
+  // handle logic to eagerly connect to the injected ethereum provider, if it exists and has granted access already
+  const triedEager = useEagerConnect();
+
+  // handle logic to connect in reaction to certain events on the injected ethereum provider, if it exists
+  useInactiveListener(!triedEager);
+
   const [open, setOpen] = useState(false);
 
   const handleClose = (connector: AbstractConnector | undefined) => {
-    if (connector) console.log(connector);
+    connector && activate(connector);
     setOpen(false);
   };
 
@@ -28,14 +43,18 @@ export default function AppBar() {
             <Typography variant="h5" component="div" sx={{ flexGrow: 1 }}>
               Arweave Social dApp
             </Typography>
-            <Button
-              variant="contained"
-              color="primary"
-              size="large"
-              onClick={() => setOpen(true)}
-            >
-              Connect
-            </Button>
+            {!!account ? (
+              account
+            ) : (
+              <Button
+                variant="contained"
+                color="primary"
+                size="large"
+                onClick={() => setOpen(true)}
+              >
+                Connect
+              </Button>
+            )}
           </Toolbar>
         </MuiAppBar>
       </Box>
