+++ constants/wallets.ts
@@ -1,6 +1,5 @@
 import { AbstractConnector } from "@web3-react/abstract-connector";
 
-import INJECTED_ICON_URL from "../assets/images/arrow-right.svg";
 import METAMASK_ICON_URL from "../assets/images/metamask.svg";
 import WALLETCONNECT_ICON_URL from "../assets/images/walletConnectIcon.svg";
 import COINBASE_ICON_URL from "../assets/images/coinbaseWalletIcon.svg";
@@ -19,15 +18,6 @@ interface WalletInfo {
 }
 
 export const SUPPORTED_WALLETS: { [key: string]: WalletInfo } = {
-  INJECTED: {
-    connector: injected,
-    name: "Injected",
-    iconURL: INJECTED_ICON_URL,
-    description: "Injected web3 provider.",
-    href: null,
-    color: "#010101",
-    primary: true,
-  },
   METAMASK: {
     connector: injected,
     name: "MetaMask",
